update and delete sauce function ok

+++ controllers/sauce.js
@@ -1,5 +1,7 @@
 const Sauce = require('../models/Sauce');
+const fs = require('fs');
 
+// CREATE A NEW SAUCE
 exports.createSauce = (req, res) => {
   // parse Json req to get a chain of string
   const sauceContent = JSON.parse(req.body.sauce);
@@ -25,29 +27,57 @@ exports.createSauce = (req, res) => {
     .catch(error => res.status(400).json({ message: 'error:' + error }));
 };
 
+// SHOW ALL THE SAUCES
 exports.getAllSauces = (req, res) => {
   Sauce.find()
     .then(sauces => res.status(200).json(sauces))
     .catch(error => res.status(400).json({ error }));
 };
 
+// SHOW THE SELECTED SAUCE
 exports.getOneSauce = (req, res) => {
   Sauce.findById(req.params.id)
     .then(sauce => res.status(200).json(sauce))
     .catch(error => res.status(400).json({ error }));
 };
 
+// MODIFY THE SAUCE
 exports.updateSauce = (req, res) => {
-  Sauce.findByIdAndUpdate(req.params.id, {
-    ...req.body,
-    _id: req.params.id,
-  })
-    .then(() => res.status(200).json({ message: 'Sauce update !' }))
-    .catch(error => res.status(400).json({ error }));
+  const sauceContent = req.file
+    ? {
+        ...JSON.parse(req.body.sauce),
+        imageUrl: `${req.protocol}://${req.get('host')}/images/${
+          req.file.filename
+        }`,
+      }
+    : { ...req.body };
+  delete sauceContent._userId;
+  Sauce.findById(req.params.id).then(sauce => {
+    if (sauce.userId != req.auth.userId) {
+      res.status(401).json({ message: 'Unauthorized' });
+    } else {
+      Sauce.findByIdAndUpdate(req.params.id, {
+        ...sauceContent,
+        _id: req.params.id,
+      })
+        .then(() => res.status(200).json({ message: 'Sauce update !' }))
+        .catch(error => res.status(401).json({ error }));
+    }
+  });
 };
 
+// DELETE THE SAUCE
 exports.deleteSauce = (req, res) => {
-  Sauce.findByIdAndDelete(req.params.id)
-    .then(() => res.status(200).json({ message: 'Sauce deleted !' }))
-    .catch(error => res.status(400).json({ error }));
+  Sauce.findById(req.params.id).then(sauce => {
+    if (sauce.userId != req.auth.userId) {
+      res.status(401).json({ message: 'Unauthorized' });
+    } else {
+      const filename = sauce.imageUrl.split('/images/')[1];
+      fs.unlink(`images/${filename}`, () => {
+        Sauce.findByIdAndDelete(req.params.id)
+          .then(() => res.status(200).json({ message: 'Sauce deleted !' }))
+          .catch(error => res.status(401).json({ error }));
+      });
+    }
+  });
 };
